third_party/blink/web_tests/external/wpt/html/webappapis/scripting/events/event-handler-attributes-body-window-expected.txt

This is a testharness.js-based test.
Found 25 FAIL, 0 TIMEOUT, 0 NOTRUN.
[FAIL] shadowed pagereveal (document.body)
  assert_equals: alternative body should reflect expected (function) function "() => 0" but got (undefined) undefined
[FAIL] shadowed pageswap (document.body)
  assert_equals: alternative body should reflect expected (function) function "() => 0" but got (undefined) undefined
[FAIL] shadowed unload (document.body)
  assert_equals: alternative body should reflect expected (function) function "() => 0" but got (object) null
[FAIL] not shadowed copy (document.body)
  assert_equals: window should reflect expected (object) null but got (undefined) undefined
[FAIL] not shadowed cut (document.body)
  assert_equals: window should reflect expected (object) null but got (undefined) undefined
[FAIL] not shadowed paste (document.body)
  assert_equals: window should reflect expected (object) null but got (undefined) undefined
[FAIL] shadowed pagereveal removal (document.body)
  assert_equals: body should reflect changes to itself expected (object) null but got (undefined) undefined
[FAIL] shadowed pageswap removal (document.body)
  assert_equals: body should reflect changes to itself expected (object) null but got (undefined) undefined
[FAIL] shadowed pagereveal on body fires when event dispatched on window
  assert_true: expected true got undefined
[FAIL] shadowed pageswap on body fires when event dispatched on window
  assert_true: expected true got undefined
[FAIL] shadowed rejectionhandled on body fires when event dispatched on window
  assert_true: expected true got undefined
[FAIL] shadowed unload on body fires when event dispatched on window
  assert_true: expected true got undefined
[FAIL] shadowed pagereveal (document.createElement("body"))
  assert_equals: body should reflect expected (function) function "() => 0" but got (undefined) undefined
[FAIL] shadowed pageswap (document.createElement("body"))
  assert_equals: body should reflect expected (function) function "() => 0" but got (undefined) undefined
[FAIL] shadowed unload (document.createElement("body"))
  assert_equals: body should reflect expected (function) function "() => 0" but got (object) null
[FAIL] not shadowed copy (document.createElement("body"))
  assert_equals: window should reflect expected (object) null but got (undefined) undefined
[FAIL] not shadowed cut (document.createElement("body"))
  assert_equals: window should reflect expected (object) null but got (undefined) undefined
[FAIL] not shadowed paste (document.createElement("body"))
  assert_equals: window should reflect expected (object) null but got (undefined) undefined
[FAIL] shadowed pagereveal removal (document.createElement("body"))
  assert_equals: alternative body should reflect changes to itself expected (object) null but got (undefined) undefined
[FAIL] shadowed pageswap removal (document.createElement("body"))
  assert_equals: alternative body should reflect changes to itself expected (object) null but got (undefined) undefined
[FAIL] shadowed pagereveal (window)
  assert_equals: body should reflect expected (function) function "() => 0" but got (undefined) undefined
[FAIL] shadowed pageswap (window)
  assert_equals: body should reflect expected (function) function "() => 0" but got (undefined) undefined
[FAIL] shadowed unload (window)
  assert_equals: body should reflect expected (function) function "() => 0" but got (object) null
[FAIL] shadowed pagereveal removal (window)
  assert_equals: body should reflect expected (object) null but got (undefined) undefined
[FAIL] shadowed pageswap removal (window)
  assert_equals: body should reflect expected (object) null but got (undefined) undefined
Harness: the test ran to completion.
